fix(lib/feedback/imu): use `f32` like Electrical

+++ lib/feedback/src/lib.rs
@@ -106,22 +106,22 @@ impl Science {
 #[repr(C)]
 pub struct Imu {
     // accel
-    pub accel_x: f64,
-    pub accel_y: f64,
-    pub accel_z: f64,
+    pub accel_x: f32,
+    pub accel_y: f32,
+    pub accel_z: f32,
 
     // gyro
-    pub gyro_x: f64,
-    pub gyro_y: f64,
-    pub gyro_z: f64,
+    pub gyro_x: f32,
+    pub gyro_y: f32,
+    pub gyro_z: f32,
 
     // compass
-    pub compass_x: f64,
-    pub compass_y: f64,
-    pub compass_z: f64,
+    pub compass_x: f32,
+    pub compass_y: f32,
+    pub compass_z: f32,
 
     // temperature
-    pub temp_c: f64,
+    pub temp_c: f32,
 }
 
 impl Imu {
